contined refactoring
